refactor: move Signal class to radar_signal.h/.cpp
