+++ src/radar_signal.h
@@ -13,11 +13,36 @@
 #include "config.h"
 #include "interpolation_point.h"
 #include "jones_vector.h"
-#include "rssignal.h"
 
 namespace rs
 {
-	class Signal; // Forward declaration
+	class Signal
+	{
+	public:
+		Signal() : _data(nullptr), _size(0), _rate(0) {}
+
+		~Signal() { delete[] _data; }
+
+		void clear();
+
+		void load(const RS_COMPLEX* inData, unsigned samples, RS_FLOAT sampleRate);
+
+		void load(const RS_FLOAT* inData, unsigned samples, RS_FLOAT sampleRate);
+
+		[[nodiscard]] RS_FLOAT rate() const { return _rate; }
+
+		[[nodiscard]] unsigned size() const { return _size; }
+
+		[[nodiscard]] RS_FLOAT* copyData() const;
+
+		std::shared_ptr<std::complex<double>[]> render(const std::vector<InterpPoint>& points,
+		                                               unsigned& size, double fracWinDelay) const;
+
+	private:
+		RS_COMPLEX* _data;
+		unsigned _size;
+		RS_FLOAT _rate;
+	};
 
 	class RadarSignal : public boost::noncopyable
 	{
